Minor tests with labels.dtype in code/train_val_test_utilities.py.

+++ code/train_val_test_utilities.py
@@ -93,7 +93,7 @@ def run_model(model, dataloader, mode, device, optimizer=None, class_specific=Tr
                 # Calculate cluster cost
                 # prototypes_of_correct_class = torch.t(model.module.prototype_class_identity[:,label]).cuda()
                 # prototypes_of_correct_class = torch.t(model.module.prototype_class_identity[:,label]).to(device)
-                labels_ = torch.reshape(labels, (-1,))
+                labels_ = torch.reshape(labels, (-1,)).to(torch.int64)
                 print(labels_.shape)
                 print(labels_.dtype)
                 print(labels_)
